+++ lib/carrierwave/base64/adapter.rb
@@ -1,7 +1,8 @@
 module Carrierwave
   module Base64
     module Adapter
-      # rubocop:disable Metrics/MethodLength
+      # rubocop:disable Metrics/MethodLength, Metrics/AbcSize
+      # rubocop:disable Metrics/CyclomaticComplexity
       def mount_base64_uploader(attribute, uploader_class, options = {})
         mount_uploader attribute, uploader_class, options
 
@@ -13,11 +14,16 @@ module Carrierwave
           return super(data) unless data.is_a?(String) &&
                                     data.strip.start_with?('data')
 
+          options[:file_name] ||= -> { attribute }
+          if options[:file_name].is_a?(Proc) && options[:file_name].arity == 1
+            options[:file_name] = options[:file_name].curry[self]
+          end
           super(Carrierwave::Base64::Base64StringIO.new(
-            data.strip, options[:file_name] || 'file'
+            data.strip, options[:file_name]
           ))
         end
-        # rubocop:enable Metrics/MethodLength
+        # rubocop:enable Metrics/MethodLength, Metrics/AbcSize
+        # rubocop:enable Metrics/CyclomaticComplexity
       end
     end
   end